fix: update test imports and mock paths after ceo.py module extraction
- Update _build_ceo_task imports from factory.cli.ceo to factory.cli._task_builder
- Update _chain_modes imports from factory.cli.ceo to factory.cli.run
- Update mock paths for _run_single_cycle and _auto_detect_mode
- Fix _get_projects_dir mock to patch at usage site (_ceo_helpers)
- Change agent role 'qa' to 'health_checker' in tests
- Update --mode qa to --mode deep-qa in TestCmdCeoQa tests
- Remove unused _resolve_model import from ceo.py
