+++ tests/test_runner.py
@@ -70,27 +70,27 @@ class TestResolvePromptWithWorkflowMode:
 
 class TestBuildCeoTaskNoSkillRead:
     def test_improve_mode_no_skill_read_instruction(self, tmp_path: Path) -> None:
-        from factory.cli.ceo import _build_ceo_task
+        from factory.cli._task_builder import _build_ceo_task
 
         task = _build_ceo_task(tmp_path, "improve")
         assert "read `skills/workflow-" not in task
         assert "playbook" in task.lower()
 
     def test_build_mode_no_skill_read_instruction(self, tmp_path: Path) -> None:
-        from factory.cli.ceo import _build_ceo_task
+        from factory.cli._task_builder import _build_ceo_task
 
         task = _build_ceo_task(tmp_path, "build")
         assert "read `skills/workflow-" not in task
 
     def test_create_mode_no_skill_read_instruction(self, tmp_path: Path) -> None:
-        from factory.cli.ceo import _build_ceo_task
+        from factory.cli._task_builder import _build_ceo_task
 
         task = _build_ceo_task(tmp_path, "create")
         assert "read `skills/workflow-" not in task
         assert "skills/workflow-create/SKILL.md" not in task
 
     def test_research_mode_no_skill_read_instruction(self, tmp_path: Path) -> None:
-        from factory.cli.ceo import _build_ceo_task
+        from factory.cli._task_builder import _build_ceo_task
 
         task = _build_ceo_task(tmp_path, "research")
         assert "read `skills/workflow-" not in task
